+++ src/Ext/Building/Body.h
@@ -28,6 +28,7 @@ public:
 	int TurretAnimFiringFrame;
 	int TurretAnimRateTick;
 	int ConstructionStartFacing;
+	bool IsPlayingRoofProductionAnim;
 
 	BuildingExt(BuildingClass* OwnerObject) : TechnoExt(OwnerObject)
 		, DeployedTechno { false }
@@ -44,6 +45,7 @@ public:
 		, TurretAnimIdleFrame { 0 }
 		, TurretAnimFiringFrame { -1 }
 		, TurretAnimRateTick { 0 }
+		, IsPlayingRoofProductionAnim { false }
 	{ }
 
 	// typed owner accessor (shadows the TechnoClass one from the base)